Reduce delta of code changes from profiler branch to master (#1537)
