+++ packages/dd-trace/src/profiling/config.js
@@ -2,6 +2,7 @@
 
 const coalesce = require('koalas')
 const os = require('os')
+const { URL } = require('url')
 const { AgentExporter } = require('./exporters/agent')
 const { FileExporter } = require('./exporters/file')
 const { ConsoleLogger } = require('./loggers/console')
@@ -14,7 +15,10 @@ const {
   DD_ENV,
   DD_TAGS,
   DD_SERVICE,
-  DD_VERSION
+  DD_VERSION,
+  DD_TRACE_AGENT_URL,
+  DD_AGENT_HOST,
+  DD_TRACE_AGENT_PORT
 } = process.env
 
 class Config {
@@ -41,9 +45,14 @@ class Config {
     this.logger = ensureLogger(options.logger)
     this.flushInterval = flushInterval
 
+    const hostname = coalesce(options.hostname, DD_AGENT_HOST, 'localhost')
+    const port = coalesce(options.port, DD_TRACE_AGENT_PORT, 8126)
+    this.url = new URL(coalesce(options.url, DD_TRACE_AGENT_URL,
+      `http://${hostname || 'localhost'}:${port || 8126}`))
+
     this.exporters = ensureExporters(options.exporters || [
-      new AgentExporter(options)
-    ], options)
+      new AgentExporter(this)
+    ], this)
 
     this.profilers = options.profilers || [
       new CpuProfiler(),
